Remove Ansible provisioning from Terraform config file.
The way we were provisioning the servers in the past was to wait for all
the servers to be up & running and finally creating a null_resource
which would invoke the calling of the Ansible playbook that will setup
the whole DCOS installation in all of the servers.
The problem was that sometimes, because of connectivity issues, some of
the playbooks would fail but Terraform would not register this, then
considering the null_resource as successfully created. Then, one would
have to manually run the Ansible playbooks, defeating the purpose of the
automatic calling from the null_resource.
So we have decided to just remove it and require from the user to simply
call the ansible playbooks by themselves once the Terraform run is over.